Codes of used 2 ESP32 modules and a ESP Cam module
Here this contains how the train automation happen with obstacle detection, energy saving mechanisms along with separately built door opening and closing mechanism. With the ESP cam module the system has provided the ability to see how the train works in real time. 
This is a prototype module with lots of extended development opportunities.